requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
nltk==3.8.1
requests==2.32.3
pytest==8.3.2
groq
numpy==1.24.3